refactor(pfb): extract table preview into FieldPreviewTable
